Make the ref input field larger
The input field is too narrow for most of the refs. The third quartile
of the line length of refs is around ~43 characters. When using the
relative unit ch[1] and a variable-width font [2] recommends setting ch
to 75% of the number of characters (43*0,75=32,25). After some trial and
error I settled on 35ch, which is enough for a ref, with 43 characters,
like org.gtk.Gtk3theme.Greybird-Geeko-Light/3.22 on my system.

[1] https://developer.mozilla.org/en-US/docs/Learn/CSS/Building_blocks/Values_and_units#relative_length_units
[2] https://meyerweb.com/eric/thoughts/2018/06/28/what-is-the-css-ch-unit/

Fix #13